# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aerobay Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/accessories": {
            "get": {
                "tags": [
                    "Accessory"
                ],
                "summary": "List of accessories",
                "operationId": "13bcb670334fde0dfb67da8da40e2d67",
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Analog FPV camera Foxeer Cat 3 Mini"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some description"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 350
                                                    },
                                                    "discount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://www.dronesdirect.co.uk/Images/CP.ZM.00000136.01_1_12679407_Supersize.jpg?v=8"
                                                    },
                                                    "dimensions": {
                                                        "type": "string",
                                                        "example": "152 x 152 x 82 mm"
                                                    },
                                                    "weight": {
                                                        "type": "string",
                                                        "example": "1200g"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "Black"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "Camera"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "manufacturer_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Accessory"
                ],
                "summary": "Create new accessory",
                "operationId": "6034a5aa16a5c9d93f4e3617e523e2f5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Analog FPV camera Foxeer Cat 3 Mini"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 350
                                            },
                                            "discount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "https://www.dronesdirect.co.uk/Images/CP.ZM.00000136.01_1_12679407_Supersize.jpg?v=8"
                                            },
                                            "dimensions": {
                                                "type": "string",
                                                "example": "152 x 152 x 82 mm"
                                            },
                                            "weight": {
                                                "type": "string",
                                                "example": "1200g"
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "Black"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "Camera"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "manufacturer_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Accessory added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Analog FPV camera Foxeer Cat 3 Mini"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some description"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 350
                                                },
                                                "discount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "https://www.dronesdirect.co.uk/Images/CP.ZM.00000136.01_1_12679407_Supersize.jpg?v=8"
                                                },
                                                "dimensions": {
                                                    "type": "string",
                                                    "example": "152 x 152 x 82 mm"
                                                },
                                                "weight": {
                                                    "type": "string",
                                                    "example": "1200g"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "Black"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Camera"
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "manufacturer_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
...
    }
}